requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
jinja2